requirements.txt
azure-functions
pillow
PyMuPDF